Provide default implementations for auto detection
