+++ chartjs-java-model/src/main/java/software/xdev/chartjs/model/options/scale/cartesian/logarithmic/LogarithmicScaleOptions.java
@@ -15,14 +15,15 @@
  */
 package software.xdev.chartjs.model.options.scale.cartesian.logarithmic;
 
-import software.xdev.chartjs.model.options.scale.cartesian.CartesianScaleOptions;
+import software.xdev.chartjs.model.options.scale.cartesian.AbstractCartesianScaleOptions;
 
 
 /**
  * @see <a href="https://www.chartjs.org/docs/latest/axes/cartesian/logarithmic.html">Chart JS</a>
  * @see <a href="https://github.com/chartjs/Chart.js/blob/v4.4.3/src/types/index.d.ts#L3312>ChartJS Source</a>
  */
-public class LogarithmicScaleOptions extends CartesianScaleOptions<LogarithmicScaleOptions, LogarithmicTickOptions>
+public class LogarithmicScaleOptions
+	extends AbstractCartesianScaleOptions<LogarithmicScaleOptions, LogarithmicTickOptions>
 {
 	public LogarithmicScaleOptions()
 	{
